+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,18 @@
 # ai
 
+## 5.0.2
+
+### Patch Changes
+
+- 401d73e: fix (ai): support dynamic tool calls in lastAssistantMessageIsCompleteWithToolCalls
+- 69fde99: Set status to ready when reconnect stream is null
+
+## 5.0.1
+
+### Patch Changes
+
+- 4d0c108: fix(ai/ui): convert provider metadata for system messages to model messages
+
 ## 5.0.0
 
 ### Major Changes