+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/azure
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [63e2016]
+  - @ai-sdk/openai@2.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [bc45e29]
+  - @ai-sdk/openai@2.0.1
+
 ## 2.0.0
 
 ### Major Changes