+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/google-vertex
 
+## 3.0.2
+
+### Patch Changes
+
+- b9cd900: feat(providers/google-vertex) Add TaskType support for Text Embedding Model
+- Updated dependencies [f5464aa]
+  - @ai-sdk/google@2.0.1
+
+## 3.0.1
+
+### Patch Changes
+
+- 11e3ba4: Make revisedPrompt nullish in schema
+
 ## 3.0.0
 
 ### Major Changes