+++ packages/togetherai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/togetherai
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [0e8ed8e]
+  - @ai-sdk/openai-compatible@1.0.1
+
 ## 1.0.0
 
 ### Major Changes