+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/xai
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [0e8ed8e]
+  - @ai-sdk/openai-compatible@1.0.1
+
 ## 2.0.0
 
 ### Major Changes